+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/fluids/FluidOffer.java
@@ -1,0 +1,28 @@
+package net.mehvahdjukaar.moonlight.api.fluids;
+
+public record FluidOffer(SoftFluidStack fluid, int minAmount) {
+    //min amount is minimum amount that offer will consume regardless of what happens if successful of course
+
+    public static FluidOffer of(SoftFluidStack stack, int min) {
+        if (stack.getCount() < min) {
+            throw new IllegalStateException("Minimum fluid amount was bigger than actual fluid amount for fluid stack "+ stack);
+        }
+        return new FluidOffer(stack, min);
+    }
+
+    public static FluidOffer of(SoftFluidStack stack) {
+        return of(stack, 1);
+    }
+
+    public static FluidOffer of(Holder<SoftFluid> fluid, int amount, int minAmount) {
+        return of(SoftFluidStack.of(fluid, amount), minAmount);
+    }
+
+    public static FluidOffer of(Holder<SoftFluid> fluid, int amount) {
+        return of(SoftFluidStack.of(fluid, amount), 1);
+    }
+
+    public static FluidOffer of(Holder<SoftFluid> fluid) {
+        return of(SoftFluidStack.of(fluid, 1), 1);
+    }
+}